convert InvokeArgumentException into runtime
The usage of `CommandLine::process` should not force the catching of an
internal exception.

An alternative would be to catch the exception in the method-call, but
this might cause issues if the user is only depending on the library to
parse setup arguments that'll be used outside of the processing.